Use pre-commit to reformat code
Use pre-commit to reformat code

+++ autonomous_driving/occupancy_prediction/projects/mmdet3d_plugin/bevformer/dense_heads/bevformer_occ_head.py
@@ -5,29 +5,19 @@
 # ---------------------------------------------
 
 import copy
+
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
-from mmcv.cnn import Linear, bias_init_with_prob
-from mmcv.utils import TORCH_VERSION, digit_version
-
-from mmdet.core import (multi_apply, multi_apply, reduce_mean)
-from mmdet.models.utils.transformer import inverse_sigmoid
+from mmcv.cnn.bricks.transformer import build_positional_encoding
+from mmcv.runner import BaseModule, auto_fp16, force_fp32
 from mmdet.models import HEADS
-from mmdet.models.dense_heads import DETRHead
-from mmdet3d.core.bbox.coders import build_bbox_coder
-from projects.mmdet3d_plugin.core.bbox.util import normalize_bbox
-from mmcv.cnn.bricks.transformer import build_positional_encoding
-from mmcv.runner import force_fp32, auto_fp16
-from projects.mmdet3d_plugin.models.utils.bricks import run_time
-import numpy as np
-import mmcv
-import cv2 as cv
-from projects.mmdet3d_plugin.models.utils.visual import save_tensor
-from mmcv.cnn.bricks.transformer import build_positional_encoding
-from mmdet.models.utils import build_transformer
 from mmdet.models.builder import build_loss
-from mmcv.runner import BaseModule, force_fp32
+from mmdet.models.dense_heads import DETRHead
+from mmdet.models.utils import build_transformer
+from mmdet.models.utils.transformer import inverse_sigmoid
+from projects.mmdet3d_plugin.models.utils.bricks import run_time
+from projects.mmdet3d_plugin.models.utils.visual import save_tensor
+
 
 @HEADS.register_module()
 class BEVFormerOccHead(BaseModule):
@@ -61,14 +51,13 @@ class BEVFormerOccHead(BaseModule):
         self.bev_h = bev_h
         self.bev_w = bev_w
         self.fp16_enabled = False
-        self.num_classes=kwargs['num_classes']
-        self.use_mask=use_mask
+        self.num_classes = kwargs['num_classes']
+        self.use_mask = use_mask
 
         self.with_box_refine = with_box_refine
         self.as_two_stage = as_two_stage
         if self.as_two_stage:
             transformer['as_two_stage'] = self.as_two_stage
-
 
         self.pc_range = pc_range
         self.real_w = self.pc_range[3] - self.pc_range[0]
@@ -151,7 +140,7 @@ class BEVFormerOccHead(BaseModule):
 
         outs = {
             'bev_embed': bev_embed,
-            'occ':occ_outs,
+            'occ': occ_outs,
         }
 
         return outs
@@ -166,25 +155,25 @@ class BEVFormerOccHead(BaseModule):
              gt_bboxes_ignore=None,
              img_metas=None):
 
-        loss_dict=dict()
-        occ=preds_dicts['occ']
-        assert voxel_semantics.min()>=0 and voxel_semantics.max()<=17
-        losses = self.loss_single(voxel_semantics,mask_camera,occ)
-        loss_dict['loss_occ']=losses
+        loss_dict = dict()
+        occ = preds_dicts['occ']
+        assert voxel_semantics.min() >= 0 and voxel_semantics.max() <= 17
+        losses = self.loss_single(voxel_semantics, mask_camera, occ)
+        loss_dict['loss_occ'] = losses
         return loss_dict
 
-    def loss_single(self,voxel_semantics,mask_camera,preds):
-        voxel_semantics=voxel_semantics.long()
+    def loss_single(self, voxel_semantics, mask_camera, preds):
+        voxel_semantics = voxel_semantics.long()
         if self.use_mask:
-            voxel_semantics=voxel_semantics.reshape(-1)
-            preds=preds.reshape(-1,self.num_classes)
-            mask_camera=mask_camera.reshape(-1)
-            num_total_samples=mask_camera.sum()
-            loss_occ=self.loss_occ(preds,voxel_semantics,mask_camera, avg_factor=num_total_samples)
+            voxel_semantics = voxel_semantics.reshape(-1)
+            preds = preds.reshape(-1, self.num_classes)
+            mask_camera = mask_camera.reshape(-1)
+            num_total_samples = mask_camera.sum()
+            loss_occ = self.loss_occ(preds, voxel_semantics, mask_camera, avg_factor=num_total_samples)
         else:
             voxel_semantics = voxel_semantics.reshape(-1)
             preds = preds.reshape(-1, self.num_classes)
-            loss_occ = self.loss_occ(preds, voxel_semantics,)
+            loss_occ = self.loss_occ(preds, voxel_semantics, )
         return loss_occ
 
     @force_fp32(apply_to=('preds'))
@@ -199,9 +188,8 @@ class BEVFormerOccHead(BaseModule):
         # return self.transformer.get_occ(
         #     preds_dicts, img_metas, rescale=rescale)
         # print(img_metas[0].keys())
-        occ_out=preds_dicts['occ']
-        occ_score=occ_out.softmax(-1)
-        occ_score=occ_score.argmax(-1)
-
+        occ_out = preds_dicts['occ']
+        occ_score = occ_out.softmax(-1)
+        occ_score = occ_score.argmax(-1)
 
         return occ_score
